+++ src/main/java/pan/artem/tinkoff/controller/error/externalservice/ExternalServiceException.java
@@ -1,0 +1,17 @@
+package pan.artem.tinkoff.controller.error.externalservice;
+
+import pan.artem.tinkoff.controller.error.MyAppException;
+
+public class ExternalServiceException extends MyAppException {
+
+    public ExternalServiceException() {
+    }
+
+    public ExternalServiceException(String message) {
+        super(message);
+    }
+
+    public ExternalServiceException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
