+++ src/main/java/pan/artem/tinkoff/controller/error/externalservice/ExternalServiceInternalErrorException.java
@@ -1,0 +1,8 @@
+package pan.artem.tinkoff.controller.error.externalservice;
+
+public class ExternalServiceInternalErrorException extends ExternalServiceException {
+
+    public ExternalServiceInternalErrorException(String message) {
+        super(message);
+    }
+}
