+++ src/main/java/pan/artem/tinkoff/controller/error/externalservice/ExternalServiceLimitExceededException.java
@@ -1,0 +1,8 @@
+package pan.artem.tinkoff.controller.error.externalservice;
+
+public class ExternalServiceLimitExceededException extends ExternalServiceException {
+
+    public ExternalServiceLimitExceededException(String message) {
+        super(message);
+    }
+}
