--- src/main/java/pan/artem/tinkoff/controller/error/weatherapi/WeatherApiInternalErrorException.java
@@ -1,8 +1,0 @@
-package pan.artem.tinkoff.controller.error.weatherapi;
-
-public class WeatherApiInternalErrorException extends WeatherApiException {
-
-    public WeatherApiInternalErrorException(String message) {
-        super(message);
-    }
-}
