--- src/main/java/pan/artem/tinkoff/controller/error/weatherapi/WeatherApiLimitExceededException.java
@@ -1,8 +1,0 @@
-package pan.artem.tinkoff.controller.error.weatherapi;
-
-public class WeatherApiLimitExceededException extends WeatherApiException {
-
-    public WeatherApiLimitExceededException(String message) {
-        super(message);
-    }
-}
