+++ src/main/java/pan/artem/tinkoff/service/CurrentWeatherServiceImpl.java
@@ -6,7 +6,7 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import io.github.resilience4j.core.exception.AcquirePermissionCancelledException;
 import io.github.resilience4j.ratelimiter.RateLimiter;
 import org.springframework.web.client.RestTemplate;
-import pan.artem.tinkoff.controller.error.weatherapi.WeatherApiException;
+import pan.artem.tinkoff.controller.error.externalservice.ExternalServiceException;
 import pan.artem.tinkoff.controller.error.RequestNotPermitted;
 import pan.artem.tinkoff.dto.WeatherDto;
 
@@ -35,7 +35,7 @@ public class CurrentWeatherServiceImpl implements CurrentWeatherService {
         } catch (io.github.resilience4j.ratelimiter.RequestNotPermitted | AcquirePermissionCancelledException e) {
             throw new RequestNotPermitted(e);
         } catch (Exception e) {
-            throw new WeatherApiException(
+            throw new ExternalServiceException(
                     "Exception occurred when executing Weather API call", e
             );
         }
@@ -47,7 +47,7 @@ public class CurrentWeatherServiceImpl implements CurrentWeatherService {
             JsonNode current = root.path("current");
             temp = current.path("temp_c").asInt();
         } catch (JsonProcessingException e) {
-            throw new WeatherApiException(
+            throw new ExternalServiceException(
                     "Exception occurred when parsing Weather API response", e
             );
         }
